Set sort_csv to use the current working dir for scratch files
GNU sort creates scratch files as part of its process. These are
normally saved in the system /tmp path. When dealing with such large
files, however, this /tmp partition is often not large enough. Using the
CWD is safer as we've probably checked we have lots of disk space in the
area in which we're working.